Implemented IPv4 class
Global variables for the class are fields that are required to be
implemented.
There are two constructors: one that has no arguments which initializes
all fields. The other accepts all fields required for implemenation as
arguments.
Finally, there are setters and getters (methods) for each field.

Class has been created without main program so it's subject to change.